requirements.txt
# Core dependencies
python>=3.11

# Network simulation
containernet>=2.3.0
mininet>=2.3.0

# Testing and performance
iperf3>=0.1.11
scapy>=2.5.0
fping>=0.0.1

# Monitoring
prometheus-client>=0.19.0
psutil>=5.9.0

# Data processing and visualization
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.17.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
myst-parser>=2.0.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
click>=8.1.0
rich>=13.5.0
loguru>=0.7.0

# Docker
docker>=6.1.0
